refactored door
